# python/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_create_file_api_fs__project_id___filepath__put": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_create_file_api_fs__project_id___filepath__put",
        "type": "object"
      },
...
      "CreateFileResponse": {
        "properties": {
          "filepath": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ResponseStatus"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "CreateFileResponse",
        "type": "object"
      },
      "DeleteFileResponse": {
        "properties": {
          "filepath": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ResponseStatus"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "DeleteFileResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResponseStatus": {
        "description": "An enumeration.",
        "enum": [
          "ok",
          "error"
        ],
        "title": "ResponseStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "RefStudio API",
    "version": "0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/fs/{project_id}/{filepath}": {
      "delete": {
        "operationId": "delete_file_api_fs__project_id___filepath__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filepath",
            "required": true,
            "schema": {
              "format": "path",
              "title": "Filepath",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteFileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete File",
        "tags": [
          "filesystem"
        ]
      },
      "get": {
        "operationId": "read_file_api_fs__project_id___filepath__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filepath",
            "required": true,
            "schema": {
              "format": "path",
              "title": "Filepath",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read File",
        "tags": [
          "filesystem"
        ]
      },
      "head": {
        "operationId": "head_file_api_fs__project_id___filepath__head",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filepath",
            "required": true,
            "schema": {
              "format": "path",
              "title": "Filepath",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Head File",
        "tags": [
          "filesystem"
        ]
      },
      "put": {
        "operationId": "create_file_api_fs__project_id___filepath__put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filepath",
            "required": true,
            "schema": {
              "format": "path",
              "title": "Filepath",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_file_api_fs__project_id___filepath__put"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateFileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create File",
        "tags": [
          "filesystem"
        ]
      }
    },
...
  }
}